% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{install_msgen}
\alias{install_msgen}
\title{Install Python Libraries}
\usage{
install_msgen(method = "pip", version = "python2.7")
}
\arguments{
\item{path}{Path to Python version under which the package should be installed.
Must be character string of the path to the version (e.g. "/usr/bin/python2.7")
(Note: must be a version of Python 2 such as 2.7)}
}
\description{
Install msgen and Azure Storage Python Libraries.
}
