% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msgen.R
\docType{package}
\name{msgen-package}
\alias{msgen-package}
\alias{msgen}
\alias{microsoft-genomics}
\title{R functions for interfacing with the Microsoft Genomics service in Azure.}
\description{
The \code{msgen} package provides funtionality  similar to the \code{msgen} Python command line interface and subsequently use the Microsoft Genomics service in Azure.
This package allows users to submit, list, check the status, and cancel workflows in the Genomics service.
}
\details{
Package: msgen\cr
Type: Package\cr
Version: 0.2.0\cr
Date: 2021-01-12\cr
License: Apache 2.0\cr
}
\author{
Colby T. Ford, Ph.D., \email{colby.ford@uncc.edu} \url{https://orcid.org/0000-0002-7859-3622}
}
\keyword{genomics,}
\keyword{microsoft}
\keyword{package,}
