% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{submit}
\alias{submit}
\title{Interact with the Microsoft Genomics service in Azure}
\usage{
submit(api_url_base, subscription_key, process_args, input_storage_account_name,
  input_storage_account_key, input_storage_account_container, blob_name_1,
  blob_name_2 = NULL, output_storage_account_name, output_storage_account_key,
  output_storage_account_container)
}
\arguments{
\item{api_url_base}{URL for your Microsoft Genomics service. (Example = `"https://eastus2.microsoftgenomics.net"`)}

\item{subscription_key}{Subscription key for your Microsoft Genomics service. (Example = "04afabfc1af94c8285faec2f15e4e459")
Go to Azure portal and open your Genomics account page. Under the `Management` heading, choose `Access keys`. There, you find both the API URL and your access keys.}

\item{process_args}{Process arguments for specifying a reference genome. Select from `"R=b37m1"`, `"R=hg38m1"`, `"R=hg38m1x"`, or `"R=hg19m1"`.}

\item{input_storage_account_name}{Azure storage account name where input files reside.}

\item{input_storage_account_key}{Azure storage account key which will be used to create temporary access tokens for input files.}

\item{input_storage_account_container}{Azure blob container where input files reside.}

\item{blob_name_1}{First file name.}

\item{blob_name_2}{Second file name, needed only if input is in the FASTQ format.}

\item{output_storage_account_name}{Azure storage account name where output files will be placed.}

\item{output_storage_account_key}{Azure storage account key which will be used to create a temporary access token for an output container.}

\item{output_storage_account_container}{Azure blob container where output files will be placed.}

\item{workflow_id}{Workflow ID returned after submitting a job.}
}
\description{
Submit, list, check the status, and cancel genomics workflows in the Microsoft Genomics service in Azure.
}
