% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_config.R
\name{use_config}
\alias{use_config}
\title{Interact with the Microsoft Genomics service in Azure}
\usage{
submit(path_to_config, task)
}
\arguments{
\item{path_to_config}{Path character string to config file.}

\item{task}{Task for the msgen system to invoke.
(Choose from "submit", "list", "status", or "cancel".)}
}
\description{
Submit, list, check the status, and cancel genomics workflows in the Microsoft Genomics service in Azure.
}
