% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{workflows}
\alias{workflows}
\alias{list_workflows}
\title{Interact with the Microsoft Genomics service in Azure}
\usage{
submit_workflow(subscription_key,
       region,
       description = "Submission from msgen R package.",
       process = "snapgatk",
       reference = "hg19m1",
       input_storage_account_name,
       input_storage_account_key,
       input_container_name,
       blob_name_1,
       blob_name_2 = NULL,
       output_storage_account_name,
       output_storage_account_key,
       output_container_name)
       
list_workflows(subscription_key,
     region)
          
get_workflow_status(subscription_key,
       region,
       workflow_id = NULL)
          
cancel_workflow(subscription_key,
       region,
       workflow_id = NULL)
}
\arguments{
\item{subscription_key}{Subscription key for your Microsoft Genomics service. (Example = "04afabfc1af94c8285faec2f15e4e459")
Go to Azure portal and open your Genomics account page. Under the `Management` heading, choose `Access keys`. There, you find both the API URL and your access keys.}

\item{region}{Region of your Microsoft Genomics service. (Example = `"eastus"`)}

\item{description}{Description string for your submission. (Default = "Submission from msgen R package.")}

\item{process}{Defines which pipeline to run. Either `"snapgatk"` (default) or `"gatk4"`.}

\item{reference}{Process arguments for specifying a reference genome. Select from `"b37m1"`, `"hg38m1"`, `"hg38m1x"`, or `"hg19m1"` (default).}

\item{input_storage_account_name}{Azure storage account name where input files reside.}

\item{input_storage_account_key}{Azure storage account key which will be used to create temporary access tokens for input files.}

\item{input_container_name}{Azure blob container where input files reside.}

\item{blob_name_1}{First file name.}

\item{blob_name_2}{Second file name, needed only if input is in the FASTQ format.}

\item{output_storage_account_name}{Azure storage account name where output files will be placed. (Optional: Will default to input storage account if not specified.)}

\item{output_storage_account_key}{Azure storage account key which will be used to create a temporary access token for an output container. (Optional: Will default to input key if not specified.)}

\item{output_container_name}{Azure blob container where output files will be placed. (Optional: Will default to input container if not specified.)}

\item{workflow_id}{Workflow ID returned after submitting a job.}
}
\description{
Submit, list, check the status, and cancel genomics workflows in the Microsoft Genomics service in Azure.
}
\examples{
submit_workflow(subscription_key = "04afabfc...",
       region = "eastus",
       process = "snapgatk",
       reference = "b37m1",
       input_storage_account_name = "mygenomicsstorage",
       input_storage_account_key= "6GyBAbvgw5sqo2...",
       input_storage_account_container = "myinputdata",
       blob_name_1 = "NA12878-chr21_1.fq.gz",
       blob_name_2 = "NA12878-chr21_2.fq.gz")
       
list_workflows(subscription_key = "04afabfc...",
     region = "eastus",)
     
get_workflow_status(subscription_key = "04afabfc...",
       region = "eastus",
       workflow_id = "12g3c5a...")

cancel_workflow(subscription_key = "04afabfc...",
       region = "eastus",
       workflow_id = "12g3c5a...")

}
